+++ src/schemas/index.ts
@@ -1,1 +1,3 @@
+export * from './director-address-schema'
 export * from './office-address-schema'
+export * from './validators'
